******* CAMBIOS 17022026 ****** 1) Corregido la descarga de fracciones y unidades 2) corregida la descarga de decimanles en unidades 3) modificada la tarjeta de inventario para mostrar los valores de unidad de medida y nombre de fraccion 4) modificada la interfza de la aplicacion

+++ app/src/main/java/com/example/acae30/modelos/Inventario.kt
@@ -8,7 +8,7 @@ data class Inventario (
     var Unidad_medida:String, // 5
     var Fraccion:Float?, // 6
     var Nombre_fraccion:String?, // 7
-    var Existencia:Int?, // 12
+    var Existencia:Float?, // 12
     var Costo:Float?, // 8
     var costo_iva:Float?, // 9
     var Precio_iva:Float?, // 17
